requirements.txt
flask>=2.3
flask-sock>=0.7
pywinpty>=2.0; sys_platform == 'win32'
